Split splash_page.dart in his own file and add his routes.dart
