+++ lib/presentation/routes.dart
@@ -3,6 +3,7 @@ import 'package:tasky/domain/entities/task.dart';
 import 'package:tasky/domain/values/unique_id.dart';
 import 'package:tasky/presentation/pages/login/login_page.dart';
 import 'package:tasky/presentation/pages/main/main_page.dart';
+import 'package:tasky/presentation/pages/splash/splash_page.dart';
 import 'package:tasky/presentation/pages/task_list/archive_page.dart';
 import 'package:tasky/presentation/pages/task_form/task_form_page.dart';
 import 'package:tasky/presentation/pages/task_info/task_info_page.dart';
@@ -13,6 +14,7 @@ class Routes {
   Routes._();
 
   /// Route names.
+  static const String splash = "/";
   static const String main = "/main";
   static const String login = "/login";
   static const String archive = "/archive";
@@ -22,6 +24,11 @@ class Routes {
   /// Route generator callback used to build the app's named routes.
   static Route<dynamic> onGenerateRoute(RouteSettings settings) {
     switch (settings.name) {
+      case Routes.splash:
+        return MaterialPageRoute(
+          builder: (context) => SplashPage(),
+          settings: settings,
+        );
       case Routes.main:
         return MaterialPageRoute(
           builder: (context) => MainPage(),
